# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "Project": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "workerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "viewerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "workers": {
            "items": {},
            "type": "array"
          },
          "viewers": {
            "items": {},
            "type": "array"
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "rateHour": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 1
          },
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "text",
          "state"
        ],
        "example": {
          "id": "c382a47e-ed48-4700-b2df-baa89aba3f22"
        }
      },
...
      "ProjectSearchSortDto": {
        "properties": {
          "createdAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "updatedAt": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "title": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "rateHour": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "trackScreenshots": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          },
          "trackProcesses": {
            "type": "string",
            "enum": [
              "ASC",
              "DESC"
            ]
          }
        },
        "type": "object"
      },
      "ProjectSearchFilterDto": {
        "properties": {
          "userId": {
            "format": "uuid",
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "type": "string"
          },
          "state": {
            "enum": [
              "Inactive",
              "Active"
            ],
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "rateHourFrom": {
            "type": "number"
          },
          "rateHourTo": {
            "type": "number"
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "withScreenshots": {
            "type": "boolean"
          },
          "withProcesses": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ProjectSearchDto": {
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/ProjectSearchSortDto",
            "type": "object"
          },
          "filter": {
            "$ref": "#/components/schemas/ProjectSearchFilterDto",
            "type": "object"
          },
          "page": {
            "minLength": 1,
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "type": "object",
        "required": [
          "sort",
          "filter",
          "page"
        ]
      },
...
      "InvoiceCreateDto": {
        "properties": {
          "fromUnix": {
            "type": "number"
          },
          "toUnix": {
            "type": "number"
          }
        },
        "type": "object",
        "required": [
          "fromUnix",
          "toUnix"
        ]
      },
      "User_search": {
        "type": "object",
        "description": "Subset of User serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "address": {
            "type": "string"
          },
          "emailOrPhone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "ROLE_USER"
              ],
              "type": "string"
            },
            "type": "array"
          }
        }
      },
...
      "Project_search": {
        "type": "object",
        "description": "Subset of Project serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "workerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "viewerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "workers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User_search"
            }
          },
          "viewers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User_search"
            }
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "rateHour": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 1
          },
          "user": {
            "$ref": "#/components/schemas/User_search"
          },
          "invoices": {
            "$ref": "#/components/schemas/Invoice_search"
          },
          "time": {
            "$ref": "#/components/schemas/Time_search"
          }
        }
      },
      "Project_create": {
        "type": "object",
        "description": "Subset of Project serialized with class-transformer group \"create\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "workerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "viewerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "rateHour": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "ProjectStatistics_search": {
        "type": "object",
        "description": "Subset of ProjectStatistics serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "processName": {
            "type": "string",
            "minLength": 1
          },
          "timeMin": {
            "type": "number"
          },
          "period": {
            "enum": [
              "1D",
              "7D",
              "1W",
              "1M",
              "6M",
              "1Y"
            ],
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Project_edit": {
        "type": "object",
        "description": "Subset of Project serialized with class-transformer group \"edit\".",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          },
          "workerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "viewerAddresses": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "trackScreenshots": {
            "type": "boolean"
          },
          "trackProcesses": {
            "type": "boolean"
          },
          "text": {
            "type": "string",
            "minLength": 1
          },
          "rateHour": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Time_search": {
        "type": "object",
        "description": "Subset of Time serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "project": {
            "$ref": "#/components/schemas/Project_search"
          },
          "user": {
            "$ref": "#/components/schemas/User_search"
          },
          "isPaid": {
            "type": "boolean"
          },
          "note": {
            "type": "string"
          },
          "screenshot": {
            "type": "string"
          },
          "processes": {
            "items": {},
            "type": "array"
          },
          "keyboardKeys": {
            "type": "number"
          },
          "minutesActive": {
            "type": "number"
          },
          "mouseKeys": {
            "type": "number"
          },
          "mouseDistance": {
            "type": "number"
          },
          "fromAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "toAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        }
      },
...
      "Invoice_search": {
        "type": "object",
        "description": "Subset of Invoice serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "project": {
            "$ref": "#/components/schemas/Project_search"
          },
          "fromAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "toAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "amount": {
            "minLength": 1,
            "type": "string"
          },
          "state": {
            "minLength": 1,
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (same value the API returns on login)"
      }
    }
  },
  "info": {
    "title": "API schema",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/project/search": {
      "post": {
        "operationId": "ProjectController.search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSearchDto",
                "type": "object",
                "required": [
                  "page",
                  "filter",
                  "sort"
                ],
                "properties": {
                  "page": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Zero-based page index"
                  },
                  "sort": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Map of field name to sort direction (ASC or DESC)"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Optional page size override"
                  },
                  "filter": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Endpoint-specific filters (see summary / examples)"
                  }
                }
              },
              "example": {
                "filter": {
                  "state": "DRAFT"
                },
                "sort": {
                  "createdAt": "DESC"
                },
                "page": 0
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/Project_search"
                        }
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  }
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Search projects accessible to the current user as owner, worker, or viewer",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/project": {
      "post": {
        "operationId": "ProjectController.create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project_create"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Create project",
        "tags": [
          "Project"
        ]
      }
    },
    "/api/project/{id}": {
      "get": {
        "operationId": "ProjectController.read",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project_search"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get project by id (project owner, workers, and viewers)",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "ProjectController.edit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project_edit"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Update project (including worker and viewer addresses)",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "ProjectController.delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Soft-delete project",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/project/{id}/stats/{period}": {
      "get": {
        "operationId": "ProjectController.getStats",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          },
          {
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectStatistics_search"
                  }
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get project statistics for project owner, workers, and viewers",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/project/{id}/close": {
      "post": {
        "operationId": "ProjectController.close",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Close project",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/invoice/project/{projectId}": {
      "post": {
        "operationId": "InvoiceController.create",
        "parameters": [
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Project"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreateDto"
              },
              "example": {
                "fromUnix": 1781301344322,
                "toUnix": 1781387744322
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Invoice_search"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "Project does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "Project does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create invoice from logged time in a range",
        "tags": [
          "Invoice"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}